backend/pyproject.toml
[tool.poetry]
name = "medicationsapi"
version = "0.1.0"
description = ""
authors = ["hasansajedi <hassansajedi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
httpx = "^0.24.1"
fastapi-pagination = "^0.12.9"
pydantic = "^2.3.0"
pandas = "^2.1.0"
python-dotenv = "^1.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
fastapi-limiter = "^0.1.5"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
pytest = "^7.4.1"
pytest-asyncio = "^0.21.1"
aioresponses = "^0.7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
